+++ src/calender/Time.cpp
@@ -10,7 +10,7 @@
 #define SECS_PER_HR (MINS_PER_HR*SECS_PER_MIN)
 
 
-bool isValidTime(uint_fast8_t hour, uint_fast8_t min, uint_fast8_t sec) {
+bool isValidTime(uint_fast16_t hour, uint_fast16_t min, uint_fast16_t sec) {
 	return (hour > 0 && hour < 24) && (min > 0 && min < 60) && (sec > 0 && sec < 60);
 }
